LPS-140640 Rename Filter Components
